boot: allow mocking of seed.ReadSystemEssential from outside of boot package
